Add preemptive recaching mechanism
Preemptive recaching is meant to solve the problem of requests dogpiling
when the http representation of an expensive and often requested
resource expires.

The problem is better described here http://stackoverflow.com/q/9858557.

The mechanism added in this commit adds two new configuration elements:
1. preemptive_recache_seconds: controls how long before the expiration
of a resource should the resource be recached (force-refreshed). While
the resource is being recached, the stale representation will keep on
being served.

2. preemptive_recache_callback: a callback function that should re-issue
the current request via some asynchronous mechanism (say, an rq job).
The request should arrive with the special header 'X-Webcache-Recache'
equal to the current resource salt.